chore(EventDelegator): move file
